license resolution now interface allowing fully custom implementations
